refactor: :recycle: move app.data.constants module to app.constants and update import statements
