+++ app/data/scripts/parse_xml_unicode_db.py
@@ -5,6 +5,7 @@ from lxml import etree
 from lxml.etree import _Element, _ElementTree
 
 from app.config.api_settings import UnicodeApiSettings
+from app.constants import NULL_BLOCK, NULL_PLANE
 from app.core.result import Result
 from app.data.constants import NULL_BLOCK, NULL_PLANE
 from app.data.encoding import get_codepoint_string
